pyproject.toml
[tool.poetry]
name = "testwork-sima-land"
version = "0.1.0"
description = ""
authors = ["Aleksey Lesnikov"]
readme = "README.md"
packages = [{include = "testwork_sima_land"}]

[tool.poetry.dependencies]
python = "^3.10"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
aiohttp = "^3.8.4"
pyyaml = "^6.0"
aiohttp-apispec = "^2.2.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.15"}
asyncpg = "^0.27.0"
greenlet = "^2.0.2"
aiohttp-session = {extras = ["secure"], version = "^2.12.0"}
pyjwt = "^2.7.0"
aiohttp-pydantic = "^1.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
